Orchestration/Workers/requirements.txt
fastapi
uvicorn
redis
celery
torch
diffusers
transformers
accelerate
faster-whisper
pillow
python-multipart
websockets